third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-not-last.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<div id="host">
    <span id="distributed">This text should be green</span>
</div>
<script>
    description("Accept simple selectors after ::content For Polymer 0.5 compat");

    var root = host.createShadowRoot();
    root.innerHTML = '<style>[select]::content.x span { color: green }</style><content class="x" select="span"></content>';

    shouldBeEqualToString("getComputedStyle(distributed).color", "rgb(0, 128, 0)");
</script>
