third_party/WebKit/LayoutTests/fast/dom/shadow/content-pseudo-element-not-last-expected.txt

CONSOLE WARNING: line 9: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Accept simple selectors after ::content For Polymer 0.5 compat

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(distributed).color is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE

This text should be green
